Fill home.js and import home in index.js
